Update README.md
Original Text:
The original weather API provider was to be the Weather Underground API, but the
free service from that API was discontinued earlier this year.
Now the direction has changed to using the Yahoo! Weather API.
Revision:
It is expected that this app will use the Yahoo! Weather API.
Rationale:
The information about not being able to use the original expected API is unnecessary for the audience, especially if the app is not even released yet.