Update README.md
Original Text:
The recommendation will draw on factors like the temperature, humidity, likelihood of rain, cloudiness, and windiness. Perhaps even non-weather factors such as elevation could be incorporated to give the best possible recommendation for the user.
Revision:
The recommendation will rely on factors such as the temperature, humidity, raininess, cloudiness, and windiness. In time, factors such as elevation could be incorporated for even better recommendation.
Rationale:
Replaced ‘like’ with more appropriate ‘such as’. Replaced ‘likelihood of rain’ with ‘raininess’. Deleted the reference to elevation being a non-weather factor since it is obvious that elevation is a non-weather factor and therefore irrelevant. Changed ‘best possible recommendation for the user’ to ‘even better recommendation’ as in prediction, the best possible recommendation is only theoretically possible.