Update README.md
Original Text:
I hope that this app will help people not make a fuss in the morning thinking about how
they will have to prepare for the weather for the day ahead. Instead, the appropriate 
clothing selection will be provided to them, and all they would have to do would be wearing
the clothing selection without wasting valuable time in the morning.
Revision:
This app will help its users save valuable time in the morning by saving them the trouble of having to figure out the appropriate clothing for the weather.
Rationale:
The revision is much more concise compared to the original. The word ‘people’ is switched to ‘users’ for less colloquial tone.